Update helpers and controllers to pass existing tests
