+++ lib/controllers/register.js
@@ -199,100 +199,103 @@ module.exports = function (req, res, next) {
       },
       'text/html': function () {
         var writeFormError = helpers.writeFormError.bind(null, req, res, view, viewModel);
+        var formActionUri = config.web.register.uri;
 
-        switch (req.method) {
-          // We should render the registration template.
-          case 'GET':
+        return helpers.handleMultiTenantRequests(req, res, formActionUri, function () {
+          switch (req.method) {
+            // We should render the registration template.
+            case 'GET':
 
-            if (helpers.requiresOrganizationResolution(config, req)) {
-              return helpers.parentDomainRedirect(req, res, config);
-            }
-
-            helpers.render(req, res, view, {
-              form: helpers.sanitizeFormData(req.body),
-              formModel: viewModel.form
-            });
-            break;
-
-          // The user is submitting a registration request, so we should attempt
-          // to validate the user's data and create their account.
-          case 'POST':
-            async.waterfall([
-              // What we'll do here is simply set default values for `givenName` and
-              // `surname`, because these value are annoying to set if you don't
-              // care about them.  Eventually Stormpath is going to remove these
-              // required fields, but for now this is a decent workaround to ensure
-              // people don't have to deal with that stuff.
-              function (callback) {
-                applyDefaultAccountFields(config, req);
-                callback();
-              },
-              function (callback) {
-                handlePreRegistration(req.body, function (err) {
-                  if (err) {
-                    return writeFormError(err);
-                  }
-
-                  helpers.validateAccount(req.body, config, function (errors) {
-                    if (errors) {
-                      return writeFormError(errors[0]);
-                    }
-
-                    callback();
-                  });
-                });
-              },
-              function (callback) {
-                helpers.prepAccountData(req.body, config, function (accountData) {
-                  accountStore.createAccount(accountData, function (err, account) {
-                    if (err) {
-                      logger.info('A user tried to create a new account, but this operation failed with an error message: ' + err.developerMessage);
-                      callback(err);
-                    } else {
-                      res.locals.user = account;
-                      req.user = account;
-                      callback(null, account);
-                    }
-                  });
-                });
-              }
-            ], function (err, account) {
-              if (err) {
-                return writeFormError(err);
+              if (helpers.requiresOrganizationResolution(config, req)) {
+                return helpers.parentDomainRedirect(req, res, config);
               }
 
-              if (account.status === 'UNVERIFIED') {
-                return handleResponse(account, defaultUnverifiedHtmlResponse);
-              }
-
-              if (config.web.register.autoLogin) {
-                var options = {
-                  username: req.body.email,
-                  password: req.body.password
-                };
-
-                return helpers.authenticate(options, req, res, function (err, expandedAccount, authResult) {
-                  if (err) {
-                    return writeFormError(err);
-                  }
-
-                  helpers.createSession(authResult, expandedAccount, req, res);
-
-                  handleResponse(expandedAccount, defaultAutoAuthorizeHtmlResponse);
-                });
-              }
-
-              helpers.expandAccount(account, config.expand, logger, function (err, expandedAccount) {
-                req.user = expandedAccount;
-
-                handleResponse(expandedAccount, defaultCreatedHtmlResponse);
+              helpers.render(req, res, view, {
+                form: helpers.sanitizeFormData(req.body),
+                formModel: viewModel.form
               });
-            });
-            break;
+              break;
 
-          default:
-            next();
-        }
+            // The user is submitting a registration request, so we should attempt
+            // to validate the user's data and create their account.
+            case 'POST':
+              async.waterfall([
+                // What we'll do here is simply set default values for `givenName` and
+                // `surname`, because these value are annoying to set if you don't
+                // care about them.  Eventually Stormpath is going to remove these
+                // required fields, but for now this is a decent workaround to ensure
+                // people don't have to deal with that stuff.
+                function (callback) {
+                  applyDefaultAccountFields(config, req);
+                  callback();
+                },
+                function (callback) {
+                  handlePreRegistration(req.body, function (err) {
+                    if (err) {
+                      return writeFormError(err);
+                    }
+
+                    helpers.validateAccount(req.body, config, function (errors) {
+                      if (errors) {
+                        return writeFormError(errors[0]);
+                      }
+
+                      callback();
+                    });
+                  });
+                },
+                function (callback) {
+                  helpers.prepAccountData(req.body, config, function (accountData) {
+                    accountStore.createAccount(accountData, function (err, account) {
+                      if (err) {
+                        logger.info('A user tried to create a new account, but this operation failed with an error message: ' + err.developerMessage);
+                        callback(err);
+                      } else {
+                        res.locals.user = account;
+                        req.user = account;
+                        callback(null, account);
+                      }
+                    });
+                  });
+                }
+              ], function (err, account) {
+                if (err) {
+                  return writeFormError(err);
+                }
+
+                if (account.status === 'UNVERIFIED') {
+                  return handleResponse(account, defaultUnverifiedHtmlResponse);
+                }
+
+                if (config.web.register.autoLogin) {
+                  var options = {
+                    username: req.body.email,
+                    password: req.body.password
+                  };
+
+                  return helpers.authenticate(options, req, res, function (err, expandedAccount, authResult) {
+                    if (err) {
+                      return writeFormError(err);
+                    }
+
+                    helpers.createSession(authResult, expandedAccount, req, res);
+
+                    handleResponse(expandedAccount, defaultAutoAuthorizeHtmlResponse);
+                  });
+                }
+
+                helpers.expandAccount(account, config.expand, logger, function (err, expandedAccount) {
+                  req.user = expandedAccount;
+
+                  handleResponse(expandedAccount, defaultCreatedHtmlResponse);
+                });
+              });
+              break;
+
+            default:
+              next();
+          }
+        });
       }
     }, next);
   });
