+++ lib/helpers/index.js
@@ -27,5 +27,8 @@ module.exports = {
   strippedAccountResponse: require('./stripped-account-response'),
   toggleMultiTenancyFields: require('./toggle-multi-tenancy-fields'),
   requiresOrganizationResolution: require('./requires-organization-resolution'),
-  parentDomainRedirect: require('./parent-domain-redirect')
+  parentDomainRedirect: require('./parent-domain-redirect'),
+  isRootDomainMultiTenantRequest: require('./is-root-domain-multi-tenant-request'),
+  redirectToOrganization: require('./redirect-to-organization'),
+  handleMultiTenantRequests: require('./handle-multi-tenant-requests')
 };
